backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.4.0
tensorflow==2.20.0
pillow>=11.3.0
numpy>=1.26.4
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
requests==2.31.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1





